refactor: rename speechballoon
