--- packages/ui/src/components/SpeechBallonContainer/index.tsx
@@ -1,1 +1,0 @@
-export { SpeechBallonContainer } from "./SpeechBallonContainer";
