+++ packages/ui/src/components/SpeechBalloonContainer/index.tsx
@@ -1,0 +1,1 @@
+export { SpeechBalloonContainer } from './SpeechBalloonContainer';
